+++ applet/src/main/java/applet/GlobalMap.java
@@ -11,7 +11,7 @@ public class GlobalMap extends GeneralMap {
 
     public void fill(short arrayIndex) {
         map_start = (short) (arrayIndex + 1);
-        while ((PSBTdata[arrayIndex + map_size] & 0xff) != 0x00 && current_key_pair < NUM_OF_KEYPAIR - 1) {
+        while ((PSBTdata[(short) (arrayIndex + map_size)] & 0xff) != 0x00 && current_key_pair < (short) (NUM_OF_KEYPAIR - 1)) {
             current_key_pair++;
             key_pairs[current_key_pair].fill((short) (arrayIndex + map_size));
